nearbyaircraft/functions/recolheVoos/requirements.txt
requests
google-cloud-bigquery
python-dateutil
